requirements.txt
Pillow
Flask
google-generativeai
langchain
langchain-groq
